Plugin docs restructure of content (#203)
# Description
I've reorganised the content of /plugins following #153. I'm interested
to see what you think!
I was very heavy handed with the use of grids for the index pages,
because they are the best I could think of. Open to other options as
always.

I also fixed a small number links on other pages outside of /plugins
because I broke a lot of links with this restructure and I wasn't
exactly sure what I caused and what was already broken - thought it was
safer to just fix them.

Here is what the homepage of plugins would look like:

![image](https://github.com/napari/docs/assets/21295664/f4d732a3-d485-4186-89e7-1f40a7d93c96)

## Type of change
- [x] Fixes or improves existing content
- [x] Adds new content page(s)

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality

## Addendum

A small number of my commits got associated with my bot account (I set
the wrong email in my git config because I deleted my config before
creating #201 🙈). I genuinely don't know how to fix that. If it is going
to be an issue I could make a new branch and squash all the commits into
one big commit from my seankmartin account.